rework GitIntegration.cs
